RM-9000 Add docs and missing ArgChecks
