+++ Remotion/Data/DomainObjects/Persistence/Configuration/DeferredStorageSettings.cs
@@ -7,16 +7,31 @@ using Remotion.Utilities;
 
 namespace Remotion.Data.DomainObjects.Persistence.Configuration
 {
+  /// <summary>
+  /// Enables configuration of the storage definition providers through IoC in a lazy manner.
+  /// Creates an instance of <see cref="Configuration.StorageSettings"/> at the first access to one of its methods
+  /// based on the supplied <see cref="IStorageSettingsFactory"/> and <see cref="IStorageObjectFactoryFactory"/>.
+  /// This <see cref="Configuration.StorageSettings"/> object is then used in every method call afterwards as well,
+  /// supplying the actual functionality.
+  /// </summary>
   [UsedImplicitly]
   [ImplementationFor(typeof(IStorageSettings), Lifetime = LifetimeKind.Singleton)]
   public class DeferredStorageSettings : IStorageSettings
   {
     private readonly IStorageSettingsFactory _storageSettingsFactory;
     private readonly IStorageObjectFactoryFactory _storageObjectFactoryFactory;
-    private IStorageSettings? _storageSettings;
 
+    private IStorageSettings? _storageSettings;
     private IStorageSettings StorageSettings => _storageSettings ??= Initialize();
 
+    /// <summary>
+    /// Creates an uninitialized instance of this class. When one of the methods on this object is called, it
+    /// becomes initialized by creating an internal <see cref="IStorageSettings"/> object.
+    /// </summary>
+    /// <param name="storageSettingsFactory">The settings factory which creates the internal <see cref="IStorageSettings"/>.</param>
+    /// <param name="storageObjectFactoryFactory">
+    /// The object factory used by the <paramref name="storageSettingsFactory"/> to create the internal <see cref="IStorageSettings"/>.
+    /// </param>
     public DeferredStorageSettings (IStorageSettingsFactory storageSettingsFactory, IStorageObjectFactoryFactory storageObjectFactoryFactory)
     {
       ArgumentUtility.CheckNotNull("storageSettingsFactory", storageSettingsFactory);
@@ -35,6 +50,7 @@ namespace Remotion.Data.DomainObjects.Persistence.Configuration
     {
       return StorageSettings.GetStorageProviderDefinition(storageGroupTypeOrNull);
     }
+
 
     public StorageProviderDefinition GetStorageProviderDefinition (string storageProviderName)
     {
